+++ lib/staticmap.dart
@@ -1,0 +1,4 @@
+library google_maps_webservice.staticmap;
+
+export './src/core.dart';
+export './src/staticmap.dart';